+++ sample/test-rules/jest.config.cjs
@@ -8,6 +8,9 @@ module.exports = {
   // Default is 5000. None of our tests take that long; fail fast.
   testTimeout: 2000,
 
+  // Without this, wasn't able to get 'exports' to work. See -> TRACK.md
+  resolver: "../hack-jest/custom-resolver.cjs",
+
   // Load docs, once at the beginning of the tests.
   globalSetup: "./setup.jest.cjs"
 };