+++ sample/test-rules/symbolsC.test.js
@@ -4,11 +4,8 @@
 import { strict as assert } from 'assert'
 import { test, expect, describe, beforeAll } from '@jest/globals'
 
-//import { dbAuth } from 'firebase-jest-testing/firestoreTestingReadOnly';
-import { dbAuth } from '../../src/firestoreTesting/readOnly.js';
-
-//import { FieldValue } from 'firebase-jest-testing/firestoreTesting';
-import { FieldValue } from '../../src/firestoreTesting/index.js';
+import { dbAuth } from 'firebase-jest-testing/firestoreTestingReadOnly';
+import { FieldValue } from 'firebase-jest-testing/firestoreTesting';
 
 const anyDate = new Date();   // a non-server date